+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Readme file updates:
   - Concept description update.
   - Excess property check according to the new features of version 0.7.0.
+- Refactoring of `defaultResultHandler` and `ResultHandler` calls in `server.ts`.
 
 ### v0.7.0
 - Zod version is v3.0.0-beta.1
@@ -33,7 +34,7 @@ const endpoint = factory.build({
     value: z.string().transform((str) => str.length)
   }),
   handler: async ({input, options}) => ({
-    transform: 'test' // => in response: { transform: 4 }
+    value: 'test' // => in response: { value: 4 }
   })
 });
 ```